csci5410-summer-23-sdp34-main/Notification_module/Leaderboard/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.8.0
google-cloud-firestore